% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asciify.R
\name{asciify}
\alias{asciify}
\title{Convert non-ASCII characters to their ASCII equivalents}
\usage{
asciify(x, verify = TRUE)
}
\arguments{
\item{x}{A character vector to be processed.}

\item{verify}{A logical value indicating whether to verify that the result is
ASCII. Defaults to \code{TRUE}. If \code{FALSE}, the function will not check that the
result is ASCII and it may return non-ASCII characters.}
}
\value{
A character vector with non-ASCII characters replaced by their ASCII
equivalents.
}
\description{
This function replaces non-ASCII characters in a string with their ASCII
equivalents. It supports a range of European non-ASCII characters, including
Icelandic, Swedish, Norwegian, Danish, Finnish, German, Estonian, Latvian,
Lithuanian, Polish, Hungarian, Slovenian, Czech, Slovak, Maltese, Romanian,
Albanian, and Croatian.
}
\examples{
asciify("Jón Þór Birgisson") # "Jon Thor Birgisson"
asciify("förståndshandikapp") # "forstandshandikapp"
asciify("Viðareiði") # "Vidareidi"
asciify("übermensch") # "uebermensch"
asciify("Jürgen Klopp") # "Juergen Klopp"
asciify("rõõmsameelsus") # "roomsameelsus"
asciify("Mężczyzna") # "Mezczyzna"
asciify("Škoda") # "Skoda"

}
