% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catty.R
\name{catty}
\alias{catty}
\title{Construct a catty vector}
\usage{
catty(x, sep = " ")
}
\arguments{
\item{x}{The \code{character} vector we want to behave in a \code{catty} manner.}

\item{sep}{The separator to output between elements of the \code{catty} vector.}
}
\value{
The \code{catty} vector.
}
\description{
\code{\link[=catty]{catty()}} constructs a \code{catty} vector (of class \code{zmisc_catty}). A  \code{catty}
vector is a \code{character} vector that uses \code{\link[=cat]{cat()}} for print output, rather
than the standard \code{print()} command. Apart from this, it behaves like a
standard character vector.

Objects of this class are convenient return values for functions that should
generally just print a desired value without the index information of a
normal print command. This is what happens if a function returning a \code{catty}
object is called interactively. However, the return value can also be
assigned to a variable for later use, making this more flexible than having
the function output the text directly.
}
\examples{
x <- catty(letters)
x
y <- catty(month.name, sep="\n")
y

}
