% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb.R
\name{cb_apply_col_spec}
\alias{cb_apply_col_spec}
\title{Apply a col_spec to a data.frame}
\usage{
cb_apply_col_spec(
  d,
  cspec,
  warn_missing_levels = TRUE,
  set_spec_attribute = FALSE
)
}
\arguments{
\item{d}{The \code{data.frame} to which the \code{col_spec} should be applied.}

\item{cspec}{The \code{col_spec} to apply.}

\item{warn_missing_levels}{Warn if any values in \code{d} don't match
the levels of the corresponding factor.}

\item{set_spec_attribute}{Should the \code{spec} attribute of the result
be set to the \code{col_spec} (for compatibility with \code{readr} behavior).}
}
\value{
A \code{data.frame} based on \code{d}, with corresponding columns
changed according to the values specified in \code{cspec}
}
\description{
\code{cb_apply_col_spec()} applies column definitions contained
in a \code{col_spec} object (typically created using \code{\link[readr:cols]{readr::cols()}}),
to the chosen columns in a \code{data.frame}.

NOTE: This is currently only implemented for column specifications
specifying conversion to a factor variable.
}
