% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voting.R
\name{check_ties}
\alias{check_ties}
\title{Check voting ties for errors}
\usage{
check_ties(score_last_in, score_first_out, ties)
}
\arguments{
\item{score_last_in}{The score of the last seat allocated}

\item{score_first_out}{The score of the first seat not allocated}

\item{ties}{Should ties cause an "\code{error}" or should the "\code{first}" be chosen}
}
\description{
Compares the voting scores of the last seat in and first seat out, for use
with seat calculation functions. The check assumes that the calling function
correctly assignes tied seats to the first listed party, and depending on the
value of the \code{ties} parameter, it can either error (clean or dirty ties if
\code{ties=="error"}), or warn (dirty ties if \code{ties=="first"}).
}
\keyword{internal}
