% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{dhondt}
\alias{dhondt}
\title{Allocate discrete seats given continuous votes}
\usage{
dhondt(votes, seats = 7, ties = c("error", "first"))
}
\arguments{
\item{votes}{A vector with the votes of different parties. The
number of parties is determined based on the length of the vector.
Votes can be given in whole or real numbers, and there is no
requirement that they sum up to 1 or 100.}

\item{seats}{An integer giving the number of seats to allocate.
Defaults to 7, which is a common size of boards and other
deliberative bodies (and a pretty nice number in general),
but one would almost always want to specify this.}

\item{ties}{A character variable specifying what to do in case of
ties. The D'Hondt method does not specify any tie handling, so
two options are allowed. Specifying \code{"error"} (the default)
results in an error if there is a tie for the marginal allocated
seat. Specifying \code{"first"} means that the party that comes
first (in the \code{votes} vector) wins. Methods such as
allocating randomly or favoring the largest party can be achived
by ordering the \code{votes} vector prior to calling the function.}
}
\value{
An integer vector with allocations of seats to parties, in
  the same order as the parties were specified in the \code{votes}
  vector. The length of the result vector always equals the length
  of the \code{votes} vector. The sum of the result vector always
  equals the value in \code{seats}. If the \code{votes} vector
  has names, the results vector has the same names.
}
\description{
Allocate a discrete number of seats (or other resources) between
parties according to a continuous number of votes (or other claims
on the resource), in a way that approximates proportionality.
Semi-proportionality is achieved with the D'Hondt method, which is
a commonly used approach in electoral allocation.

Other implementations of D'Hondt allocation in \code{R} can be found
in \code{electoral::seats_ha()} and \code{coalitions::dHondt()}. The
motivation behind this implementation is to have a simple interface
with reasonable defaults, suitable for use in pipes operating on
data in data.frames or tibbles. Party names are also optional,
which is useful if there are no natural names for the parties.
}
\examples{
dhondt(c(13,11,7,5,3))
dhondt(c(liberals=4876, conservatives=5212, greens=2349), seats=21)

}
