% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_spec.R
\name{factor_spec}
\alias{factor_spec}
\title{Construct code to use as col_spec for factors}
\usage{
factor_spec(d)
}
\arguments{
\item{d}{The data frame to use when generating the col_spec}
}
\value{
A string with definitions suitable for defining a col_spec,
after any needed editing to reorder factors.
}
\description{
This function examines all character columns in a data.frame (typically one that
was read from a csv file using \code{readdr::read_csv()}, and generates
a specification suitable for reading those columns from the underlying file
as factors.

The key benefit is that it will find all variables that are using the same
factor levels and group them together, so editing the col_spec to reorder
factor levels or make other changes is straightforward.

Because the string contains quotation marks, it can be useful to pipe the
output directly into cat(), and then copy the output into a source file
before using it. The output can then be edited before the colspec
is used to read the data in fresh from the CSV file. Assuming the \code{cols()}
result is assigned to a variable \code{cspec}, one might have the following.\preformatted{d <- read_csv(csv_file)
factor_spec(d) |> cat()
# edit the output to fit ones needs
cspec <- cols(...)
d <- read_csv(csv_file, col_types=cspec)
}
}
