% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zuludrake.R
\name{get_drake_target_size}
\alias{get_drake_target_size}
\title{Return the size of a drake target}
\usage{
get_drake_target_size(target)
}
\arguments{
\item{target}{Name of the target}
}
\value{
Size of the target
}
\description{
This function returns the size of a given drake target.
}
