% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zuludrake.R
\name{get_drake_target_sizes}
\alias{get_drake_target_sizes}
\title{Return the size of all targets in a drake plan}
\usage{
get_drake_target_sizes(my_plan)
}
\arguments{
\item{my_plan}{Name of the plan from which to get target names}
}
\value{
Tibble containing names and sizes of the targets
}
\description{
This function returns the size of all targets
in a given drake plan
}
