% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{identity_ellipsis}
\alias{identity_ellipsis}
\alias{noop}
\title{Return argument unchanged}
\usage{
identity_ellipsis(x, ...)

noop(x, ...)
}
\arguments{
\item{x}{The single argument to (not) process.}

\item{...}{Any other arguments are ignored}
}
\value{
The argument \code{x} is returned unchanged.
}
\description{
It can often be useful (when developing pipelines for example), to have
access to a function that simply returns the element that it receives. The
\code{\link[base:identity]{base::identity()}} function works well for only one argument, but in some
cases following arguments should be ignored. In those cases the
\code{identity_ellipsis()} function can come in handy.

The function was previously called \code{noop()} but that name is now deprecated.
}
