% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enframed_and_labelled.R
\name{lookup}
\alias{lookup}
\title{Lookup values from a lookup table}
\usage{
lookup(x, lookup_table)
}
\arguments{
\item{x}{A string vector whose elements shall be looked up}

\item{lookup_table}{The lookup table to use. If the table is in the form of
a data.frame, the lookup columns should be named
\code{name} (for the key) and \code{value} (for the value).
If the lookup table is in the form of a named vector or list,
the name is used for the key, and the returned value is
taken from the values in the vector or list.}
}
\value{
A string vector based on \code{x}, with values replaced
with the lookup values from \code{lookup_table}. Any values
not found in the lookup table are returned unchanged.
}
\description{
This function implements lookup of certain strings (such as
variable names) from an lookup table which maps keys
onto values (such as variable descriptions). Original values
are returned if they are not found in the lookup table.
}
\details{
The lookup table can be in the form of an enframed data.frame,
in the form of a named vector, or in the form of a list.

Any names in x are not included in the result.
}
