% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{lookup_enframed}
\alias{lookup_enframed}
\title{This function implements lookup of certain
strings (such as variable names) from an enframe-type tibble
which maps keys onto values (such as variable descriptions).}
\usage{
lookup_enframed(x, d.enframed)
}
\arguments{
\item{d.enframed}{data.frame with two columns called "name" and "value"}
}
\description{
The value added by this function is as follows:
 - All names are cleared from the result
 - Original values are returned if not found in lookup.table
}
