% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{lpad}
\alias{lpad}
\title{Left pad string with whitespace or other characters}
\usage{
lpad(string, width, pad = " ")
}
\description{
Left pad a string with whitespace or other characters
in order to get a string of a certain length.
This function is a thin wrapper around stringr::str_pad()
}
