% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zuluverse.R
\name{move}
\alias{move}
\title{Move column or selection of columns}
\usage{
move(data, cols, ref, side = c("before", "after"))
}
\arguments{
\item{data}{A \code{data.frame}}

\item{cols}{unquoted column name or numeric or selection of columns using a select helper}

\item{ref}{unquoted column name}

\item{side}{\code{"before"} or \code{"after"}}
}
\value{
A data.frame with reordered columns
}
\description{
Column(s) described by \code{cols} are moved before (default) or after the reference
  column described by \code{ref}
}
\details{
All credit to Moody Mudskipper:
https://stackoverflow.com/questions/52096919/move-a-column-conveniently/52096938#52096938

See also:
https://github.com/tidyverse/dplyr/issues/2047
}
\examples{
iris2 <- head(iris,2)
move(iris2, Species, Sepal.Width)
move(iris2, Species, Sepal.Width, "after")
move(iris2, 5, 2)
move(iris2, 4:5, 2)
move(iris2, one_of("Sepal.Width","Species"), Sepal.Width)
move(iris2, starts_with("Petal"), Sepal.Width)

}
\seealso{
Other zuluverse: 
\code{\link{mutate_cond}()}
}
\concept{zuluverse}
