% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{na.replace}
\alias{na.replace}
\title{Replace missing values}
\usage{
na.replace(x, replace)
}
\description{
Usage: na.replace(x, replace)
Arguments
}
\details{
x : vector possibly contining missing (NA) values.
replace : scalar replacement value
}
