% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{na_if_in}
\alias{na_if_in}
\title{Replace values in x that are contained in y with NA.}
\usage{
na_if_in(x, y)
}
\arguments{
\item{x}{Vector to modify.}

\item{y}{Value to replace with \code{NA}.}
}
\value{
A modified version of \code{x} that replaces any values that are found in \code{y} with \code{NA}.
}
\description{
This function is a variant/extension of \code{\link[dplyr:na_if]{dplyr::na_if()}}. Whereas that
version requires \code{y} to be of length one, this function allows \code{y} to be a
vector and replaces any elements of \code{x} that are found in \code{y} with \code{NA}
values.

This function can be very useful when cleaning data and
setting multiple any annoying values to \code{NA}.
}
\seealso{
Other functions for \code{NA} handling: 
\code{\link{na_replace}()}
}
\concept{functions for \code{NA} handling}
