% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{na_replace}
\alias{na_replace}
\title{Replace missing values}
\usage{
na_replace(x, replace)
}
\arguments{
\item{x}{Vector possibly containing missing (\code{NA}) values.}

\item{replace}{Scalar replacement value.}
}
\value{
Copy of x with any missing values replaced with replace.
}
\description{
Replaces NA values found in a vector with a specified scalar value

\verb{NOTE:} The \code{\link[dplyr:coalesce]{dplyr::coalesce()}} function now provides this functionality in a
more general way, so this function may merit deprecation.
}
\seealso{
Other functions for \code{NA} handling: 
\code{\link{na_if_in}()}
}
\concept{functions for \code{NA} handling}
