% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{noop}
\alias{noop}
\title{Do return a single argument unchanged}
\usage{
noop(x)
}
\description{
It can be useful (when developing pipelines for example),
to have acces to a function that simply returns the elemtent
that it receives. This is that function
}
