% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notate.R
\name{notate}
\alias{notate}
\alias{ViewNotated}
\alias{ViewAnnotated}
\title{Embed factor levels and value labels in values.}
\usage{
notate(d)

ViewNotated(d)

ViewAnnotated(d)
}
\arguments{
\item{d}{The \code{data.frame} that one desires to annotate and/or view.}
}
\value{
The processed \code{data.frame}, suitable for viewing.
}
\description{
This function adds level/label information as an annotation to either factors
or \code{labelled} variables. This function is called \code{notate()} rather than
\code{annotate()} to avoid conflict with \code{\link[ggplot2:annotate]{ggplot2::annotate()}}.

When printing \code{labelled} variables from a \code{tibble} in a console, both the
numeric value and the text label are shown, but no variable labels. When
using the \code{View()} function, only variable labels are shown but no value
labels. For factors, there is no way to view the integer levels and values at
the same time.

In order to allow the viewing of both variable and value labels at the same
time, this function converts both \code{factor} and \code{labelled} variables to
\code{character}, including both numeric levels (\code{labelled} values) and character
values (\code{labelled} labels) in the output.

The output is not intended for further processing, only for viewing, for
example using the \code{View()} function. This can be achieved directly by using
the \code{ViewNotated()} or \code{ViewAnnotated()} functions (\code{ViewAnnotated()} is a
synonym for \code{ViewNotated()}.
}
