% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primes.R
\name{nth_prime}
\alias{nth_prime}
\title{Calculate and return the nth prime number}
\usage{
nth_prime(n)
}
\arguments{
\item{n}{the position of the prime number to be calculated}
}
\value{
the nth prime number
}
\description{
This function will use the \code{primes} package unless it either is missing, or
if options("zulutils.zerodeps") has been explicitly set to false, in which
case it will use a local implementation of the sieve of Erastosthenes.
}
\examples{
nth_prime(1) # 2
nth_prime(10) # 29
nth_prime(100) # 541
nth_prime(1000) # 7919
nth_prime(10000) # 104729

}
