% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{padding}
\alias{padding}
\alias{lpad}
\alias{rpad}
\title{Pad string with space or other characters}
\usage{
lpad(string, width, pad = " ")

rpad(string, width, pad = " ")
}
\arguments{
\item{string}{Character vector to pad}

\item{width}{How many characters at minimum in the output}

\item{pad}{The string to pad with}
}
\description{
Left or right pad a string with space or other
characters in order to get a string of a certain length.
This function is a thin wrapper around stringr::str_pad()
}
