% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catty.R
\name{print.zmisc_catty}
\alias{print.zmisc_catty}
\title{Print a catty vector}
\usage{
\method{print}{zmisc_catty}(x, ...)
}
\arguments{
\item{x}{The catty vector}

\item{...}{Any other arguments to print (they will be ignored)}
}
\value{
The input is returned invisibly.
}
\description{
\code{print.zmisc_catty(x, ...)} prints a \code{catty} vector (using \code{cat()} instead of
\code{print()} for the output).
}
\keyword{internal}
