% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{rename_enframed}
\alias{rename_enframed}
\title{Rename tibble columns according to values in another tibble.}
\usage{
rename_enframed(d, d.enframed)
}
\description{
Renames variables from d according to the columns
name (rename from) and value (rename to) in d.framed.
}
\details{
The expectation is that d.framed has been created
with enframe(), such as
d.framed = enframe(c(froma="To A",fromb="To B"))
}
