% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{sample_strings}
\alias{sample_strings}
\title{Sample a set of strings, each string of a given length}
\usage{
sample_strings(nchar = 3, size = 1, upper = FALSE)
}
\arguments{
\item{nchar}{The length of each string.}

\item{size}{The number of strings.}

\item{upper}{Should the strings be upper case?}
}
\value{
A random vector of \code{size} strings of length \code{nchar}.
}
\description{
Sample a set of strings, each string of a given length
}
\examples{
sample_strings()
d <- data.frame(key = sample_strings(7,9), value=1:9)
d[order(d$key),]

}
