% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{se}
\alias{se}
\title{Calculate standard error (of the mean estimate) of a vector}
\usage{
se(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector, or of a type allowable in the \code{sd()} function.}

\item{na.rm}{A logical indicating whether missing values should be removed.}
}
\value{
The standard error of mean estimate of x
}
\description{
creating the need for this function.
}
\concept{Statistical functions}
