% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primes.R
\name{sieve_primes}
\alias{sieve_primes}
\title{Find all primes up to a given positive integer}
\usage{
sieve_primes(n)
}
\arguments{
\item{n}{A positive integer specifying the upper limit for the prime numbers.}
}
\value{
A numeric vector of prime numbers up to \code{n}.
}
\description{
This function implements the Sieve of Eratosthenes algorithm to find all
prime numbers up to the given positive integer \code{n}. It returns a vector of
integers representing the prime numbers.
}
\examples{
zulutils:::sieve_primes(10)
zulutils:::sieve_primes(20)

}
\keyword{internal}
