% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{strh}
\alias{strh}
\title{A modified version of the str() function that limits the length of displayed lists.}
\usage{
strh(object, max.level = 1, list.len = nth_prime(5 - max.level), ...)
}
\arguments{
\item{object}{the object to be printed.}

\item{max.level}{the maximum number of nested levels to be printed. Default is 1.}

\item{list.len}{the maximum number of elements to be printed for a list.
Default is determined by the value of max.level and the nth prime number.}

\item{...}{additional arguments to be passed to str().}
}
\value{
A printed representation of the object.
}
\description{
A modified version of the str() function that limits the length of displayed lists.
}
