% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yencode.R
\name{suppress_warnings}
\alias{suppress_warnings}
\title{Suppress warnings that match specific regular expressions}
\usage{
suppress_warnings(expr, regexp)
}
\description{
Based on \href{https://stackoverflow.com/questions/16517795/selective-suppresswarnings-that-filters-by-regular-expression/55182432#55182432}{this stack overflow answer},
except it does not allow function arguments, only regular expressions. This removes the dependency on non-base-R packages.
}
\keyword{internal}
