% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb.R
\name{type_convert_with_labels}
\alias{type_convert_with_labels}
\title{Convert column types according to a col_spec}
\usage{
type_convert_with_labels(df, col_types = NULL, ...)
}
\arguments{
\item{df}{The \code{data.frame} to process.}

\item{col_types}{The \code{col_spec} to apply.}

\item{...}{Other arguments to \code{readr::type_convert()}.}
}
\value{
A \code{data.frame} based on \code{d}, with any factor columns specified in
\code{col_types} converted to \code{factor} \emph{and} any other columns processed
according to the functionality of \code{readr::type_convert()}
}
\description{
Custom wrapper around \code{readr::type_convert()} that suppresses warnings and
copies labels, in order to (mostly) match the behavior of
\code{cb_apply_col_spec()}. but without reimplementing a lot of functionality.
}
\details{
A known key difference compared \code{cb_apply_col_spec()} is that \code{df} is fully
processed according to the \code{col_types} argument, meaning that other columns
may be converted in addition to the factor columns.
}
