% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catty.R
\name{validate.zmisc_catty}
\alias{validate.zmisc_catty}
\title{Internal validation function for catty objects}
\usage{
validate.zmisc_catty(x)
}
\arguments{
\item{x}{The object to validate}
}
\value{
The object
}
\description{
Internal validation function for catty objects
}
\keyword{internal}
