% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{zeq}
\alias{zeq}
\title{Generate sequence in a safe way}
\usage{
zeq(from, to)
}
\arguments{
\item{from}{The lower bound of the sequence}

\item{to}{The higher bound of the sequence}
}
\value{
A sequence ranging from \code{from} to \code{to}
}
\description{
This function creates an increasing integer sequence, but differs from
the standard one in that it will not silently generate a decreasing
sequence when the second argument is smaller than the first.
If the second arg is one smaller than the first it will generate
an empty sequence, if the difference is greater, the function will
error out.
}
