% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zfit.R
\name{zglm}
\alias{zglm}
\title{Run a glm model in a pipe (see \code{zlm})}
\usage{
zglm(data, formula, ...)
}
\description{
Run a glm model in a pipe (see \code{zlm})
}
\seealso{
Other zfit: 
\code{\link{zlm}()},
\code{\link{zlogit}()},
\code{\link{zprint}()},
\code{\link{zprobit}()}
}
\concept{zfit}
