% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zfit.R
\name{zlm}
\alias{zlm}
\title{Run an lm model in a pipe}
\usage{
zlm(data, formula, ...)
}
\arguments{
\item{data}{A \code{data.frame}}

\item{...}{Other arguments to be passed to \code{lm}}
}
\value{
A fitted model
}
\description{
This function wraps around the \code{lm} function in order to make it
more friendly to pipe syntax (with the data first)
}
\examples{
cars \%>\% zlm( speed ~ dist )

iris \%>\%
  filter(Species=="setosa") \%>\%
  zlm(Sepal.Length ~ Sepal.Width + Petal.Width)

}
