% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zfit.R
\name{zlogit}
\alias{zlogit}
\title{Run a logit model in a pipe (see \code{zlm})}
\usage{
zlogit(data, formula, ...)
}
\description{
Run a logit model in a pipe (see \code{zlm})
}
\seealso{
Other zfit: 
\code{\link{zglm}()},
\code{\link{zlm}()},
\code{\link{zprint}()},
\code{\link{zprobit}()}
}
\concept{zfit}
