% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zfit.R
\name{zprobit}
\alias{zprobit}
\title{Run a probit model in a pipe (see \code{zlm})}
\usage{
zprobit(data, formula, ...)
}
\description{
Run a probit model in a pipe (see \code{zlm})
}
\seealso{
Other zfit: 
\code{\link{zglm}()},
\code{\link{zlm}()},
\code{\link{zlogit}()},
\code{\link{zprint}()}
}
\concept{zfit}
