% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zulu_letters}
\alias{zulu_letters}
\title{A lookup table for the phonetic alphabet}
\format{
A \code{tibble} (data frame) with 26 rows and 2 variables:
\describe{
\item{name}{letters of the alphabet}
\item{value}{phonetic representation of the letters}
}
}
\usage{
zulu_letters
}
\description{
A dataset containing the letters of the alphabet and their
representaion in the phonetic alphabet (alpha, bravo, charlie,
..., zulu). The data is formatted as an \verb{enframed tibble}
(see \code{\link[tibble:enframe]{tibble::enframe()}}), suitable for use in contexts such
as the \code{\link[=lookup]{lookup()}} function.
}
\keyword{datasets}
