% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_zulutils.R, R/zuluverse.R
\name{zulutils-defunct}
\alias{zulutils-defunct}
\alias{move}
\title{Defunct functions in package \pkg{zulutils}.}
\usage{
move(...)
}
\description{
The functions listed below are defunct. When possible,
  alternative functions with similar functionality are mentioned.
}
\section{\code{move}}{

\code{move()} is defunct. Use \code{\link[dplyr:relocate]{dplyr::relocate()}}.
}

\keyword{internal}
