% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_zulutils.R, R/base.R,
%   R/enframed_and_labelled.R
\name{zulutils-deprecated}
\alias{zulutils-deprecated}
\alias{na.replace}
\alias{lookup_enframed}
\title{Deprecated functions in package \pkg{zulutils}.}
\usage{
na.replace(x, replace)

lookup_enframed(x, lookup_table)
}
\description{
The functions listed below are deprecated and will be defunct
  in the near future. When possible, alternative functions with similar
  functionality are also mentioned. Help pages for deprecated functions
  are available at \code{help("<function>-deprecated")}.
}
\section{\code{na.replace}}{

\code{na.replace()} is deprecated in favor of \code{\link[dplyr:coalesce]{dplyr::coalesce()}}.
}

\section{\code{lookup_enframed}}{

\code{lookup_enframed()} is deprecated in favor of \code{\link[zmisc:lookup]{zmisc::lookup()}}.
}

\keyword{internal}
