% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{bgrep}
\alias{bgrep}
\title{Boolean grep function}
\usage{
bgrep(string, pattern)
}
\arguments{
\item{string}{The (possibly vectorized) string to process.}

\item{pattern}{The (possibly vectorized) regexp pattern to use.}
}
\value{
A vector with boolean values indicating in which elements
of \code{string} matched the corresponding elements in \code{pattern}.
}
\description{
This function looks for pattern in each of the elements of string, returning
TRUE for each element that contains pattern, and FALSE for the other
elements. This function is a thin wrapper around \code{\link[stringr:str_detect]{stringr::str_detect()}}
}
