% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{bgrep}
\alias{bgrep}
\title{Boolean grep function}
\usage{
bgrep(string, pattern)
}
\description{
This function looks for pattern in each of the elements
of string, returning TRUE for each element that contains
pattern, and FALSE for the other elements. The argument
order follows the conventions of the stringr package,
and the underlying matching is done using stringr::str_extract()
}
