% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{get_labels_enframed}
\alias{get_labels_enframed}
\title{Get all the variable labels of a labelled tibble in an enframed tibble}
\usage{
get_labels_enframed(d)
}
\arguments{
\item{d}{The tibble whose columns have variable labels}
}
\value{
A tibble in enframed format containing variable names and labels
}
\description{
Given a data tibble/data.frame, this function will return a tibble in
enframed format (two columns named name and value), where the name
column contains the names of the variables in the original tibble and
the value column contains the labels. This can be used to review the
labels, and (if coupled with a set_labels_enframed function) could be
used to preserve labels across label-destructive operations, or export
labels to a file (or a sheet in an excel file). Finally, it could be
used with rename_enframed() in a pipe to replace the names of variables
with their labels for display in a plot or table function that does not
support the display of labels in a simple way.
}
