% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{lookup_enframed}
\alias{lookup_enframed}
\title{Lookup values from a two-column tibble}
\usage{
lookup_enframed(x, d.enframed)
}
\arguments{
\item{x}{A string vector whose elements shall be looked up}

\item{d.enframed}{The (enframed) tibble to use as a lookup table}
}
\value{
A string vector based on \code{x}, with values replaced
                  with the lookup values from \code{d.enframed}. Any values
                  not found in the lookup table are returned unchanged.
}
\description{
This function implements lookup of certain strings (such as
variable names) from an (enframed) tibble which maps keys
onto values (such as variable descriptions). Original values
are returned if they are not found in the lookup tibble.
Any names in x are not included in the result.
}
