% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zuluverse.R
\name{mutate_cond}
\alias{mutate_cond}
\title{Mutate only rows of data.frame/tibble that satisfy condition}
\usage{
mutate_cond(.data, condition, ..., envir = parent.frame())
}
\description{
https://stackoverflow.com/questions/34096162/dplyr-mutate-replace-on-a-subset-of-rows
}
