% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{na_if_in}
\alias{na_if_in}
\title{Replace values in x that are contained in y with NA}
\usage{
na_if_in(x, y)
}
\description{
This function is a variant/extension of dplyr::na_if()
}
\seealso{
Other na_functions: 
\code{\link{na_replace}()}
}
\concept{na_functions}
