% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{na_replace}
\alias{na_replace}
\title{Replace missing values}
\usage{
na_replace(x, replace)
}
\arguments{
\item{x}{Vector possibly contining missing (NA) values.}

\item{replace}{Scalar replacement value}
}
\value{
Copy of x with any missing values replaced with replace
}
\description{
Replace missing values
}
\seealso{
Other na_functions: 
\code{\link{na_if_in}()}
}
\concept{na_functions}
