% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enframed_and_labelled.R
\name{rename_enframed}
\alias{rename_enframed}
\title{Rename tibble columns according to values in another tibble.}
\usage{
rename_enframed(d, d.enframed)
}
\arguments{
\item{d}{The tibble whose columns we want to rename}

\item{d.enframed}{The (enframed) tibble with the old and new names}
}
\value{
A tibble that is identical to \code{d} but with renamed columns
}
\description{
Renames variables from d according to the columns
name (rename from) and value (rename to) in d.enframed.
The expectation is that d.framed has been created
with \code{enframe()}, such as
\code{d.framed = enframe(c(a="New A", b="New B"))}.
}
