% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{zample}
\alias{zample}
\title{Sample from a vector or data.frame in a safe way}
\usage{
zample(x, size = length(x), replace = FALSE, prob = NULL)
}
\description{
This function duplicates the functionality of sample(), with the exception
that it does not attempt the (rather silly in my opinion) user-friendliness
of switching the interpretation of the first element to a number if the
length of the vector is 1.
}
\details{
To sample from an interval between 1 and n, simply use safer.sample(1:n)
(or use sample(n) instead)
}
