% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{zingle}
\alias{zingle}
\title{Return the single (unique) value found in a vector}
\usage{
zingle(x, na.rm = FALSE)
}
\arguments{
\item{x}{Vector which chould only include identical values.}

\item{na.rm}{Will NAs be removed before checking. (defaults to FALSE)}
}
\value{
The single unique element in \code{x}, or error if all elements are not identical.
}
\description{
Returns the first element in a vector, but only if all the other
elements are identical to the first one (the vector only has a
zingle value). If the elements are not all identical, it throws
an error. The vector must contain at least one non-na value,
or the function errors out as well.
}
\details{
This is a useful function in aggregations when
all values in a given group should be identical, but
you want to make sure.

Optionally takes a \code{na.rm} parameter, similarly to \code{sum}, \code{mean} and
other aggregate functions. If \code{na.rm} is \code{TRUE}, a vector with a single
unique value but some \code{NA} values mixed in is OK, because the \code{NAs} will
be removed before checks are performed.
}
