% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zulutils.R
\docType{package}
\name{zulutils}
\alias{zulutils}
\title{zulutils: Some functions that the author found useful}
\description{
This package could just as well have been named yamisc
(Yet Another MISCellaneous package).
It contains various miscellaneous utility functions that the
author found useful at the time. You are free to use it, but
YMMV (Your Mileage May Vary).
}
