% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_zulutils.R
\docType{package}
\name{zulutils}
\alias{zulutils}
\alias{zulutils-package}
\title{zulutils: Miscellaneous utility functions that the author found useful}
\description{
This package contains various miscellaneous utility
functions  that the author found useful at the time. You are
free to use  it, but YMMV (Your Mileage May Vary).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://torfason.github.io/zulutils}
  \item \url{https://github.com/torfason/zulutils}
  \item \url{https://torfason.github.io/zulutils/}
}

}
\keyword{internal}
